+++ AffiseAttributionLib/Classes/events/EventsManager.swift
@@ -8,6 +8,7 @@ import Foundation
  */
 class EventsManager {
     
+    static let TIME_SEND_SECOND: TimeInterval = 3
     static let TIME_SEND_REPEAT: TimeInterval = 15
     
     static let SCHEDULER_SEND_REPEAT: TimeInterval = 15
@@ -38,6 +39,7 @@ class EventsManager {
      * Timer fo repeat send events
      */
     private var timer: ScheduledTimer?
+    private var timerSecond: ScheduledTimer?
     
     private var scheduler: ScheduledTimer?
 
@@ -53,6 +55,9 @@ class EventsManager {
     private func subscribeToActivityEvents() {
         //Send events on started
         sendEvents(withDelay: false)
+
+        //Send events on delay
+        sendEventsDelay(delay: EventsManager.TIME_SEND_SECOND)
         
         //Start timer fo repeat send events
         startTimer()
@@ -88,6 +93,13 @@ class EventsManager {
             sendDataToServerUseCase.send(withDelay: withDelay, sendEmpty: sendEmpty)
         }
     }
+
+    private func sendEventsDelay(delay: TimeInterval) { 
+        timerSecond?.stop()
+        timerSecond = ScheduledTimer(interval: delay, repeats: false) { [weak self] in
+            self?.sendEvents(withDelay: false)
+        }
+    }
     
     /**
      * Send events (if present) on schedule in case network failures
@@ -121,6 +133,9 @@ class EventsManager {
     private func stopTimer() {
         //Stop timer
         timer?.stop()
+
+        //Stop second timer
+        timerSecond?.stop()
     }
 
     @objc